% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmlewbel.R
\name{gmmlewbel}
\alias{gmmlewbel}
\title{GMM estimator of structural equations under heteroskedastic or non-normal
errors.}
\usage{
gmmlewbel(formula, data, init = NULL)
}
\arguments{
\item{formula}{a multipart formula like y2 ~ x1 + x2 | y1 | z. See details.}

\item{data}{a data.frame or object convertable to a matrix.}

\item{init}{a vector of initial parameters. Estimated by OLS if not provided.}
}
\value{
a gmm object.
}
\description{
GMM estimator of structural equations under heteroskedastic or non-normal
errors.
}
