% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_formulas}
\alias{list_formulas}
\title{Process the formula into parts we can use.}
\usage{
list_formulas(formula, env)
}
\value{
a list of formulas.
}
\description{
Process the formula into parts we can use.
}
