% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moment_tri}
\alias{moment_tri}
\title{Moment function for a triangular system with no instruments.}
\usage{
moment_tri(nX, nZ)
}
\description{
Moment function for a triangular system with no instruments.
}
