% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moments_hetero}
\alias{moments_hetero}
\title{Moment function for heteroskedastic errors.}
\usage{
moments_hetero(nX, nZ1, nZ2, is_simultaneous)
}
\description{
Moment function for heteroskedastic errors.
}
