% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moments_nonnormal}
\alias{moments_nonnormal}
\title{Moment function for a triangular system with no instruments.}
\usage{
moments_nonnormal(nX, nZ, overid)
}
\description{
Moment function for a triangular system with no instruments.
}
